rename `multi_control_multi_target_pauli` -> `multi_control_pauli`
...
