+++ qualtran/bloqs/qubitization/qubitization_walk_operator_test.py
@@ -19,7 +19,7 @@ import pytest
 from qualtran import Adjoint
 from qualtran._infra.gate_with_registers import get_named_qubits
 from qualtran.bloqs.chemistry.ising.walk_operator import get_walk_operator_for_1d_ising_model
-from qualtran.bloqs.mcmt.multi_control_multi_target_pauli import MultiControlPauli
+from qualtran.bloqs.mcmt import MultiControlPauli
 from qualtran.bloqs.multiplexers.select_pauli_lcu import SelectPauliLCU
 from qualtran.bloqs.qubitization.qubitization_walk_operator import (
     _thc_walk_op,
